requirements.txt
flask>=2.0.0
pyserial>=3.5
mysql-connector-python>=8.0.0
pytz
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
mediapipe>=0.8.9
pillow>=8.0.0
werkzeug>=2.0.0
flask-sqlalchemy>=2.5.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
requests>=2.26.0
flask-cors>=3.0.0
gunicorn>=20.1.0
eventlet>=0.33.0
python-engineio>=4.0.0
python-socketio>=5.0.0
pytest>=6.2.0
black>=21.5b2
flake8>=3.9.0
mypy>=0.910